UIandSlotIssue_Fixed

+++ Models/WriteDTO/tbl_ProjectDetails.cs
@@ -10,11 +10,11 @@ namespace WorkStatus.Models.WriteDTO
   public class tbl_ProjectDetails
     {
         [DbColumn(IsIdentity = true, IsPrimary = true)]
-        public long SNo { get; set; }
+        public int SNo { get; set; }
         [DbColumn]
-        public long OrganizationId { get; set; }
+        public int OrganizationId { get; set; }
         [DbColumn]
-        public long ProjectId { get; set; }
+        public int ProjectId { get; set; }
         [DbColumn]
         public string TotalWorkedHours { get; set; }
         [DbColumn]
